+++ src/Oriflame.PolicyBuilder.Xml/Definitions/Common/ForwardRequest.cs
@@ -1,26 +1,13 @@
 ﻿using System;
+using System.Collections.Generic;
 using Oriflame.PolicyBuilder.Policies.Extensions;
-using Oriflame.PolicyBuilder.Xml.Mappers;
 
 namespace Oriflame.PolicyBuilder.Xml.Definitions.Common
 {
     public class ForwardRequest : PolicyXmlBase
     {
-        public ForwardRequest(TimeSpan? timeout) : this(timeout?.GetSeconds())
+        public ForwardRequest(IDictionary<string, string> attributes) : base("forward-request", attributes)
         {
-        }
-        
-        public ForwardRequest(string timeoutValue, bool? bufferResponse = null) : base("forward-request")
-        {
-            if (!string.IsNullOrEmpty(timeoutValue))
-            {
-                Attributes.Add("timeout", timeoutValue);
-            }
-
-            if (bufferResponse.HasValue)
-            {
-                Attributes.Add("buffer-response", BoolMapper.Map(bufferResponse.Value));
-            }
         }
     }
 }