+++ ChangeLog.md
@@ -1,8 +1,25 @@
 # Change log
 
+## 4.2.0
+
+`ForwardRequest` policy - added optional attribute builder - allow to set `follow-redirects`, `buffer-request-body`, `buffer-response`, `fail-on-error-status-code`, `timeout` optional attributes
+
 ## 4.1.0
 
 - `ForwardRequest` policy - allows to set `buffer-response` optional attribute
+
+## 4.0.0
+
+- New policies `DeleteHeader` and `DeleteQueryParameter`
+- `ExistAction.Delete` marked as `Obsolete` - replaced with new policies `DeleteHeader` and `DeleteQueryParameter`
+
+## 3.6.0
+
+- `ContextVariable` extended with `GetValueOrDefault` method
+
+## 3.5.0
+
+- added new attribute `CachingType` to `CacheStoreValue` and `CacheLookupValue` policies
 
 ## 3.4.0